Fix issue with annotation retention writing
As of now, if an annotation class in Kotlin sources is meta-annotated
(for example with just another manually created annotation or something
like @RequiresOptIn or @RestrictTo) and doesn't have an explicit
retention annotation, retention will not get written to signature files.
This is a problem because we want to add a compatibility check for
annotation retention (to make sure it doesn't incompatibly change
between versions), but if we try to read in retention from a signature
file that doesn't have retention written to it, Metalava will attempt
to guess the default retention based on the language (Class for Java
and Runtime for Kotlin), and crash because of unkown language type.

For this reason, we want to explicitly mark retention for as many
annotations as possible.

Note: this seems to be a problem just in Kotlin sources, Metalava
seems to write default retention to signature files when reading Java
sources.

Bug: 465178983
Bug: 202967265
Test: `./gradlew`
Change-Id: I1394e1a2913bd474bd96032c2bc8f2ebc4dc5e2d