[test refactor] Remove codename from DocAnalyzerTest
This CL:
* Remove all use of ARG_CURRENT_CODE name from the test args
* Removes all mention of code name from test cases
* Updated test behaviour with actual behaviour when codename is remvoed

Context:
--current-codename is deprecated and will be removed in later CLs in
favour of --api-version-for-sources.
It is important to understand that the @apiSince label currently relies
on the codename and whether the version is less than the current
version. This must be changed to rely on the --api-version-for-sources
only (as --current-version is also being deprecated).

This testcase has been left incomplete to display the actual behaviour
of the code. Later CLs will be added to update the tests to show what is
expected given --api-version-for-sources

Bug: 454050901
Test: ./gradlew

Change-Id: I00066588238814c9f106250d14129cdcfd68c5c3