Add test for usage of expect/actual typealias from classpath
When an expect class/actual typealias is used from a common module, with K1 the psi for the usages will be replaced with the actual typealias value, but it will not be for K2. See the linked bug for more context. There's a workaround in place for this issue when the expect/actual typealias is from sources, but not when it is from the classpath. This adds a test demonstrating the issue when the typealias is from the classpath.

Bug: 396445075
Test: UastTestBase
Change-Id: I4ab8dc4f63e877337f1835a06816427f9de64419