Use @RequiresCapabilities in CommonPackageItemTest
Adds `@RequiresCapabilities` to those tests that only work on a limited
set of providers.

Uses `HIDDEN_ITEMS` instead of `PACKAGE_HTML_FILES` on the `@hide` test
as it is the hidden aspect that does not work in snapshot. Snapshots do
support `package.html` files as long as the information has been
extracted and stored in the `PackageItem` object.

Uses `JAVA` for every test that uses `package.html` or
`package-info.java` files as they will not do anything on the text
model provider.

Bug: 461491536
Test: ./gradlew
Change-Id: Icdd9b1da7b1f7c9e801c57d8828cdcee774087fb