Improve CommonFlagTest.`Test empty flags`
Previously, it did not check the `Item.showability` which meant that it
did not detect that it worked differently on snapshots. This change
adds that check and requires `Capability.API_VARIANT_SELECTORS` as
otherwise the test would fail on snapshot.

Preparation for a follow-up change that will report issues when
attempting to revert an API without being given a previously released
API.

Bug: 454856811
Test: ./gradlew
Change-Id: Id9b7249c1517b7edd569daaa2f5109721257e813