Fix broken test with undefined Function class
The `Check generic type signature insertion` test was broken on Turbine
because the `Function` type was not defined. The issue was being
reported by the test but the report is being ignored. That will be
fixed in a follow-up change. This change fixes the issue by using the
`java.lang.function.Consumer` type instead.

Bug: 467355045
Test: ./gradlew
Change-Id: I2469c4f4360d6cd2de2dc1a49f8642ebf20b7130