Separate option validation in Options.updateClassPath
Previously, it would check that both `--sdk-home` and `--jdk-home` had
not both been specified while updating the class path. That complicated
the logic. This change separates the validation out to make it simpler
to refactor in follow-up changes.

This does slightly change the semantics as previously `--sdk-home` was
ignored if `--compile-sdk-version` was not already set but no uses of
Metalava seem to rely on that. If they do then fixing it will just
require tweaking the command line.

Bug: 467355045
Test: ./gradlew
Change-Id: I4ac7d421ecdd9f9215765437f4111de6ff347fef