Refactor Options.updateClassPath()
Previously, it would modify `mutableClassPath` directly. This changes
it to take a `List<File>` and if necessary, return a new list with
additional jars depending on the jdkHome and sdkHome properties.
Slightly less efficient but simplifies the follow-up change which
converts `--classpath` to Clikt which does not return a mutable list.

Bug: 467355045
Test: ./gradlew
Change-Id: Id0465fff63dc58ab0779c35f082b4bb3640bb84d