Remove requirement that type parameter substitutions must be ReferenceTypeItems
Previously, the `TypeItem.convertType` method was only used to map between type parameters and type arguments of a superclass/subclass pair. In this case, it was safe to assume that the type arguments were all `ReferenceTypeItem`s because it is not possible to use a wildcard as a type argument in an extends/implements clause.

However, this method is also being used for mapping between type parameters of a typealias and the type arguments in the usage of a typealias. In this case, it is possible to use a wildcard (such as `*`), which means that casting to a `ReferenceTypeItem` is not safe.

Bug: 467693145
Test: UastTestBase -- the added test fails with a ClassCastException without this change
Change-Id: Ie36386a54c60d5bf49420d28be83718e83f33f4d