Move Options.sources to SourceOptions.sourceFiles
Some of the files in `sources` come from the `--source-files` option
and some are supplied as arguments. `SourceOptions` does not process
the arguments so it needs to get them from somewhere else. This change
adds an `additionalSourceFilesProvider` parameter to `SourceOptions`
to get the files collated by `Options.parse(...)`.

A follow-up change will stop collating the files in
`Options.parse(...)` and just provide the remaining arguments collated
by `MainCommand`.

Bug: 467355045
Test: ./gradlew
Change-Id: I2eb0e464bbdcc78b82ee8a71a095904eaa921644