Limit the scope of classpath searches for KMP projects
A multiplatform project has a classpath for each source set. It is possible for conflicting definitions of the same class to be present in different source set classpaths, for instance if there are separate android and jvm source sets and a dependency has separate android and jvm compilations. In this case, when metalava processes a single jvm or android compilation, it should limit the scope of searches of the classpath to just the classpath for the main source set.

This uses the `analysisScope` provided by the analysis API for the main source set that is being analyzed to limit the scope of a classpath search for a KMP project.

Bug: 467689000
Test: UastTestBase -- without this change, the added test is flaky in whether the android or jvm classpath class gets resolved

Change-Id: If724f9ea03a3435909eb1d7703142dce42686dbb