Stop parsing arguments in Options.parse(...)
Previously, `MainCommand` would collate arguments and unknown options
in `flags` and pass them through to `Options.parse(..)` for processing.
However, it no longer does any custom processing and is just
replicating the default behavior of a `CliktCommand` that has been
temporarily disabled (by setting `treatUnknownOptionsAsArgs` to `true`.

This change removes that from `Options.parse(...)` (and renames it to
`initialize(...)`) and does that work in `MainCommand`.

Bug: 467355045
Test: ./gradlew
Change-Id: If7f6593735b245fcecba17edfcf3937b3c4e92e6