Remove ignore test as it is no longer relevant
Previously, the test tried to compare compatibility between APIs loaded
from jars. It was marked `@Ignore` because it did not work, producing
an OOM error. The underlying cause was fixed a long time ago but
uncommenting the test revealed that it is not needed as it is not
possible to check compatibility against a jar file. That is tested
elsewhere.

Bug: 308154836
Test: ./gradlew
Change-Id: I03a9c01cf86d12861d9213d5746210b994c9964d