Make Driver extension functions proper members
Previously, the package level functions in `Driver.kt` were changed to
extension functions on `Driver` as part of the work to clean it up.
This change completes that clean up by making those extension functions
proper members of `Driver`. That allowed the `Driver` properties to
have their access more tightly restricted.

This change is large but it just involves:
* Removing `Driver.` from extension functions.
* Including the functions in the body of `Driver`.
* Indenting and reformatting.

Bug: 467355045
Test: ./gradlew
Change-Id: I3c911c13ed2b0b148576f8e881a5a443a75c8491