Avoid initialization error for class FileFormat.Version
Change https://r.android.com/3884861 accidentally broke the
`HelpCommandTest`. It is not exactly clear why but after that change
`HelpCommandTest` breaks when ran on its own. It broke because it hit
an initialization cycle between `FileFormat` and `FileFormat.Version`.
Prior to that change something cause `FileFormat` to always be
initialized first which avoided the issue. After that change,
`FileFormat` was not always initialized before the
`signatureFileFormatsHelp()` method was called. That in turn called
`FileFormat.Version.entries` which tried to initialize `Version`
first causing an error.

This change avoids that by adding a `FileFormat.versions` property.

Bug: 467355045
Test: ./gradlew
Change-Id: I276a8c58ee7d286d083c3cdbaa59d04c36180778