Use DocContentPredicate to find imported name usage
Previously, `SourceFileUtils.filterImports(...)` would search
`ItemDocumentation.text` for any words that matched imported names.
This change replaces that with a `DocContentPredicate` that has the
same behavior even though that behavior results in including more
imports than necessary.

Follow-up changes will tweak the behavior to avoid importing
unnecessary imports as contrary to the comment in `filterImports(...)`
including unnecessary imports can cause problems as it can result in
the stubs not compiling.

Bug: 447588621
Test: ./gradlew
Change-Id: I078e42c0310dfafb5767f36f190395c8c385b35f