Ignore imported names in @throws reference when filtering imports
Previously, using an imported name in `@throws` reference would result
in the import being kept in the stubs just in case doclava needed it to
resolve the `@throws` reference. However, `@throws` references are
fully resolved by Metalava so this change ignores them when determining
if an import should be kept.

Bug: 447588621
Test: ./gradlew
Change-Id: Ie7ce118c759af5b73750c9e462ab2698a94efbeb