Make fullyQualifiedDocumentation(String) an implementation detail
Previously, `fullyQualifiedDocumentation(String)` was defined on
`ItemDocumentation`. That is because it was needed by
`ItemDocumentation.fullyQualifiedDocumentation()` which was used by a
test. This change replaces the call to `fullyQualifiedDocumentation()`
in the test with a check of the printed output (which already fully
qualifies the documentation) and removes
`fullyQualifiedDocumentation()`. It then pushes the
`fullyQualifiedDocumentation(String)` method down as an implementation
detail of `AbstractItemDocumentation`.

Bug: 447588621
Test: ./gradlew
Change-Id: I740f1be42463ecf4003caf34316d3e2f6f5abb53