Ignore imported names matching parameter names when filtering imports
Previously, using an imported name in `@param` named would result in
the import being kept in the stubs just in case doclava needed it to
resolve the `@param` name. However, `@param` names are never resolved
against the imports so this change ignores them when determining if an
import should be kept.

Bug: 447588621
Test: ./gradlew
Change-Id: I1adf6910f071aaa98c5a698c63a0151ab5690389