Resolve class references in @see block tag
Previously, Metalava would not resolve references in `@see` block tags.
This change customizes the code that handles `{@link}` to handle `@see`
adding in special handling of the "literal string" and
`<a href="...">...</a>` forms of `@see`.

One slight difference of this over the Psi handling is that Psi treats
`@link` and `@see` differently. For `@link` it uses the original text
as the label if it resolves the reference to a different string. Psi
does not do that for `@see`. This change causes them to be both
treated the same as `@link`.

This fixes a number of tests that were failing in Turbine.

Bug: 447588621
Test: ./gradlew
Change-Id: Ic493788e037b800d85dd389c1529e13f05939304