Ignore imported names in text when filtering imports
Previously, using an imported name in text content would result in the
import being kept in the stubs just in case doclava needed it to
resolve the name. However, names in text are never resolved against the
imports so this change ignores them when determining if an import
should be kept.

Bug: 447588621
Test: ./gradlew
Change-Id: I259abbda89985c82914e2067cf6b5d081fc9cf6e