Add BaseTagTypes.registerDefaultTagType(String)
Previously, `BlockTagTypes.DEPRECATED` and `InlineTagTypes.INHERIT_DOC`
used `tagTypeOf(...)` to retrieve the `TagType`. That is the same
method as is used for retrieving `TagType`s found in the javadoc and it
will silently create and register the tag type if it could not be
found. However, follow-up changes need to provide additional
information when registering those types, i.e. that they are block and
inline respectively.

This change adds a new `registerDefaultTagType(String)` to use instead
in preparation for those changes.

Bug: 469362503
Test: ./gradlew
Change-Id: I7d3df2f4866a4eb07e6f881d77c0287c771c8b50