Force parsing of descriptions in BaseDocCommentTest
Previously, `DocComment.description` and `BlockTagSection.description`s
were parsed lazily which means that sometimes a BaseDocCommentTest test
did not report issues when creating the `DocComment` and it was up to
the individual tests to check them explicitly.

This change updates `createTestDocComment(...)` to force the
`*.description`s to be parsed reporting any issues that are found.

Bug: 469362503
Test: ./gradlew
Change-Id: I6c36fe3bb7cf28979a0f02b5ba72cab839f93202