Add test showing KMP lint issue with elements of internal class
If an element of an internal class does not have an explicit visibility modifier, it is technically considered public since this is the default kotlin visibility. However, it is effectively internal since it can only be used where the containing class can be used. This adds a test showing a lint issue when one version of an item has an explicit internal modifier and another doesn't.

Bug: 407736078
Test: MultiplatformLintTest
Change-Id: Ia7156356fbb1782c10877db38f29e14125116692