Use effective visibility in KMP lint
If an element of an internal class does not have an explicit visibility modifier, it is technically considered public since this is the default kotlin visibility. However, it is effectively internal since it can only be used where the containing class can be used. This updates the KMP lint check for visibility matches to consider effective visibility.

Bug: 407736078
Test: MultiplatformLintTest
Change-Id: Ia1c9fbb7712a8834868930fe47f4dd8a51c0a50b