Demonstrate issue with sealed class compatibility
This commit adds a test to demonstrate how the current
compatibility checks don't raise issues when an abstract
method is added to a non-final effectively sealed class.

However, compatibility errors are raised when methods are
added to sealed interfaces, even if those interfaces are
effectively final.

The followup commit will fix these problems and the tests.

Bug: 220960090
Test: `./gradlew`
Change-Id: Id2229f7a89da6e9c799a51a1b503d08a7d0f2d25